+++ src/lib.rs
@@ -38,6 +38,32 @@ fn f32_stats(v: &[f32]) -> (f32, f32, f32, f32, usize) {
 
 //--------------------------------------------------------------------------------------------------
 
+const LETTER_GRADES: [(f32, char); 5] = [
+    (90.0, 'A'),
+    (80.0, 'B'),
+    (70.0, 'C'),
+    (60.0, 'D'),
+    (0.0, 'F'),
+];
+
+/**
+Get the letter grade for a percentage score
+*/
+fn letter_grade(pct: f32) -> char {
+    LETTER_GRADES
+        .iter()
+        .find_map(|(threshold, letter)| {
+            if pct >= *threshold {
+                Some(*letter)
+            } else {
+                None
+            }
+        })
+        .unwrap()
+}
+
+//--------------------------------------------------------------------------------------------------
+
 /**
 Question bank
 */
@@ -148,8 +174,8 @@ impl Bank {
     /**
     Generate a quiz
     */
-    pub fn quiz(&self) -> Quiz {
-        Quiz::new(&self.questions)
+    pub fn quiz(&self, shuffle: bool) -> Quiz {
+        Quiz::new(&self.questions, shuffle)
     }
 }
 
@@ -194,7 +220,6 @@ impl Question {
                 _ => {}
             }
         }
-        answers.shuffle(&mut thread_rng());
         Question { content, answers }
     }
 }
@@ -245,13 +270,17 @@ impl Quiz {
     /**
     Create a new [`Quiz`]
     */
-    fn new(questions: &[Question]) -> Quiz {
+    fn new(questions: &[Question], shuffle: bool) -> Quiz {
         let mut questions = questions.to_vec();
-        let mut rng = thread_rng();
-        questions.shuffle(&mut rng);
-        questions
-            .iter_mut()
-            .for_each(|x| x.answers.shuffle(&mut rng));
+
+        if shuffle {
+            let mut rng = thread_rng();
+            questions.shuffle(&mut rng);
+            questions
+                .iter_mut()
+                .for_each(|x| x.answers.shuffle(&mut rng));
+        }
+
         Quiz { questions }
     }
 
@@ -285,10 +314,33 @@ impl Quiz {
     }
 
     /**
-    Generate quiz [`Answers`] and quiz with answers markdown
+    Generate quiz [`Answers`]
     */
-    pub fn answers(&self) -> (Answers, String) {
-        let a: BTreeMap<usize, Vec<String>> = self
+    pub fn answers(&self) -> Answers {
+        Answers::new(self)
+    }
+}
+
+//--------------------------------------------------------------------------------------------------
+
+/**
+Quiz answers
+
+This struct is used in two ways:
+
+1. Generated via [`Quiz.answers()`] during quiz generation in order to generate
+   content for the `answers.json` and `answers.md` files
+2. Loaded from a saved `answers.json` file for grading a quiz
+*/
+#[derive(Debug)]
+pub struct Answers {
+    answers: BTreeMap<usize, Vec<String>>,
+    markdown: Option<String>,
+}
+
+impl Answers {
+    fn new(quiz: &Quiz) -> Answers {
+        let answers: BTreeMap<usize, Vec<String>> = quiz
             .questions
             .iter()
             .enumerate()
@@ -311,62 +363,59 @@ impl Quiz {
             })
             .collect();
 
-        let md = self
-            .questions
-            .iter()
-            .enumerate()
-            .map(|(i, x)| {
-                let mut c = answer_counter();
-                let n = i + 1;
-                let pre = format!("{n}. ");
-                let sep = format!("\n\n{}", " ".repeat(pre.len()));
-                let ans: HashSet<_> = a.get(&n).unwrap().iter().collect();
-                format!(
-                    "{pre}{}\n\n{}",
-                    x.content
-                        .iter()
-                        .map(|x| x.replace('\n', &sep[1..]))
-                        .collect::<Vec<_>>()
-                        .join(&sep),
-                    x.answers
-                        .iter()
-                        .map(|x| {
-                            let letter = c.next().unwrap();
-                            if ans.contains(&letter) {
-                                format!("    * **{letter}) {}**\n\n", x.content)
-                            } else {
-                                format!("    * {letter}) {}\n\n", x.content)
-                            }
-                        })
-                        .collect::<Vec<_>>()
-                        .join("")
-                )
-            })
-            .collect::<Vec<_>>()
-            .join("");
+        let markdown = Some(
+            quiz.questions
+                .iter()
+                .enumerate()
+                .map(|(i, x)| {
+                    let mut c = answer_counter();
+                    let n = i + 1;
+                    let pre = format!("{n}. ");
+                    let sep = format!("\n\n{}", " ".repeat(pre.len()));
+                    let ans: HashSet<_> = answers.get(&n).unwrap().iter().collect();
+                    format!(
+                        "{pre}{}\n\n{}",
+                        x.content
+                            .iter()
+                            .map(|x| x.replace('\n', &sep[1..]))
+                            .collect::<Vec<_>>()
+                            .join(&sep),
+                        x.answers
+                            .iter()
+                            .map(|x| {
+                                let letter = c.next().unwrap();
+                                if ans.contains(&letter) {
+                                    format!("    * **{letter}) {}**\n\n", x.content)
+                                } else {
+                                    format!("    * {letter}) {}\n\n", x.content)
+                                }
+                            })
+                            .collect::<Vec<_>>()
+                            .join(""),
+                    )
+                })
+                .collect::<Vec<_>>()
+                .join(""),
+        );
 
-        (Answers { answers: a }, md)
+        Answers { answers, markdown }
     }
-}
 
-//--------------------------------------------------------------------------------------------------
-
-/**
-Quiz answers
-*/
-#[derive(Debug, Default, Deserialize, Clone)]
-pub struct Answers {
-    answers: BTreeMap<usize, Vec<String>>,
-}
-
-impl Answers {
     /**
     Load [`Answers`] from a JSON file
     */
     pub fn from(path: &Path) -> Result<Answers> {
-        let answers: BTreeMap<usize, Vec<String>> =
-            serde_json::from_str(&std::fs::read_to_string(path)?)?;
-        Ok(Answers { answers })
+        Ok(Answers {
+            answers: serde_json::from_str(&std::fs::read_to_string(path)?)?,
+            markdown: None,
+        })
+    }
+
+    /**
+    Serialize to a JSON string
+    */
+    pub fn json(&self) -> String {
+        serde_json::to_string(&self.answers).unwrap()
     }
 
     /**
@@ -377,9 +426,9 @@ impl Answers {
     }
 
     /**
-    Calculate the number of problems on the quiz
+    Return the number of questions on the quiz
     */
-    fn problems(&self) -> usize {
+    fn questions(&self) -> usize {
         self.answers.len()
     }
 
@@ -389,11 +438,12 @@ impl Answers {
     fn get(&self, key: &usize) -> Option<&Vec<String>> {
         self.answers.get(key)
     }
-}
 
-impl std::fmt::Display for Answers {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        writeln!(f, "{}", serde_json::to_string(&self.answers).unwrap())
+    /**
+    Return the Markdown content
+    */
+    pub fn markdown(&self) -> &Option<String> {
+        &self.markdown
     }
 }
 
@@ -411,7 +461,7 @@ pub struct Class {
     total: usize,
 
     #[serde(skip)]
-    problems: usize,
+    questions: usize,
 
     #[serde(skip)]
     scores: BTreeMap<String, (usize, Vec<usize>)>,
@@ -431,7 +481,7 @@ impl Class {
     */
     pub fn grade(&mut self, answers: &Answers) {
         self.total = answers.total();
-        self.problems = answers.problems();
+        self.questions = answers.questions();
         self.scores = BTreeMap::new();
         for (name, quiz) in &self.students {
             let mut missed = 0;
@@ -477,10 +527,12 @@ impl Class {
     pub fn markdown(&self) -> String {
         let mut scores: BTreeMap<usize, Vec<&str>> = BTreeMap::new();
         let mut scores_sum = 0;
-        let mut wrongs: BTreeMap<&str, String> = BTreeMap::new();
+        let mut wrongs: BTreeMap<&str, &[usize]> = BTreeMap::new();
         for (name, (score, wrong)) in &self.scores {
             wrongs.insert(
                 name,
+                wrong,
+                /*
                 if wrong.is_empty() {
                     String::from("none")
                 } else {
@@ -490,6 +542,7 @@ impl Class {
                         .collect::<Vec<_>>()
                         .join(", ")
                 },
+                */
             );
             if let Some(students) = scores.get_mut(score) {
                 students.push(name);
@@ -499,70 +552,216 @@ impl Class {
             scores_sum += score;
         }
 
-        let mut r = vec![format!(
-            "\
-# {}
-
-Name | Score | Percentage | Grade | Wrong
------|------:|-----------:|-------|-------
-\
-            ",
-            self.description,
-        )];
-
         let mut pcts = vec![];
-        let mut grades: BTreeMap<_, _> = ["A", "B", "C", "D", "F"]
-            .iter()
-            .map(|x| (x.to_string(), 0))
-            .collect();
-
+        let mut grades = Grades::new();
         for (score, students) in scores.iter().rev() {
             let pct = (*score as f32) / (self.total as f32) * 100.0;
-            let grade = if pct >= 90.0 {
-                "A"
-            } else if pct >= 80.0 {
-                "B"
-            } else if pct >= 70.0 {
-                "C"
-            } else if pct >= 60.0 {
-                "D"
-            } else {
-                "F"
-            };
-            grades.entry(grade.to_string()).and_modify(|x| *x += 1);
+            let grade = letter_grade(pct);
             for name in students {
-                r.push(format!(
-                    "{name} | {score} | {pct:.1}% | {grade} | {}\n",
+                grades.push(Grade::new(
+                    name,
+                    *score,
+                    pct,
+                    grade,
                     wrongs.get(name).unwrap(),
                 ));
                 pcts.push(pct);
             }
         }
 
+        let mut stats = Stats::new();
         let scores_keys = scores.keys().collect::<Vec<_>>();
         let (min_pct, max_pct, mean_pct, _sum, count) = f32_stats(&pcts);
         let mean_score = (scores_sum as f32) / (count as f32);
-        r.push(format!(
-            "\n\
-Description        | Value
--------------------|------------
-Number of problems | {}
-Total points       | {}
-High score         | {} ({max_pct:.1}%)
-Low score          | {} ({min_pct:.1}%)
-Mean score         | {mean_score:.1} ({mean_pct:.1}%)
-\
-            ",
-            self.problems,
-            self.total,
-            scores_keys.last().unwrap(),
-            scores_keys.first().unwrap(),
+        stats.push(Stat::new(
+            "Number of questions",
+            &self.questions.to_string(),
         ));
-        for (k, v) in &grades {
-            r.push(format!("{k}                  | {v}\n"));
+        stats.push(Stat::new("Total points", &self.total.to_string()));
+        stats.push(Stat::new(
+            "High score",
+            &format!(
+                "{}, {max_pct:.1}%, {}",
+                scores_keys.last().unwrap(),
+                letter_grade(max_pct),
+            ),
+        ));
+        stats.push(Stat::new(
+            "Low score",
+            &format!(
+                "{}, {min_pct:.1}%, {}",
+                scores_keys.first().unwrap(),
+                letter_grade(min_pct),
+            ),
+        ));
+        stats.push(Stat::new(
+            "Mean score",
+            &format!("{mean_score}, {mean_pct:.1}%, {}", letter_grade(mean_pct)),
+        ));
+        for (letter, count) in grades.histogram() {
+            stats.push(Stat::new(&letter.to_string(), &count.to_string()));
         }
-        r.push(String::new());
 
-        r.join("")
+        format!(
+            "# {}\n\n{}\n\n{}\n\n",
+            self.description,
+            grades.markdown(),
+            stats.markdown()
+        )
     }
+}
+
+//--------------------------------------------------------------------------------------------------
+
+struct Grade {
+    name: String,
+    score: usize,
+    pct: f32,
+    grade: char,
+    wrong: Vec<usize>,
+}
+
+impl Grade {
+    fn new(name: &str, score: usize, pct: f32, grade: char, wrong: &[usize]) -> Grade {
+        Grade {
+            name: name.to_string(),
+            score,
+            pct,
+            grade,
+            wrong: wrong.to_vec(),
+        }
+    }
+
+    fn markdown(&self) -> Vec<String> {
+        vec![
+            self.name.clone(),
+            self.score.to_string(),
+            format!("{:.1}%", self.pct),
+            self.grade.to_string(),
+            self.wrong
+                .iter()
+                .map(|x| x.to_string())
+                .collect::<Vec<_>>()
+                .join(", "),
+        ]
+    }
+}
+
+struct Grades {
+    grades: Vec<Grade>,
+}
+
+impl Grades {
+    fn new() -> Grades {
+        Grades { grades: vec![] }
+    }
+
+    fn push(&mut self, grade: Grade) {
+        self.grades.push(grade);
+    }
+
+    fn markdown(&self) -> String {
+        let mut rows = vec!["Name,Score,Percent,Grade,Wrong"
+            .split(',')
+            .map(|x| x.to_string())
+            .collect::<Vec<_>>()];
+        rows.append(&mut self.grades.iter().map(|x| x.markdown()).collect::<Vec<_>>());
+        md_table(&rows, &[1, 2])
+    }
+
+    fn histogram(&self) -> BTreeMap<char, usize> {
+        let mut r = LETTER_GRADES
+            .iter()
+            .map(|(_, x)| (*x, 0))
+            .collect::<BTreeMap<char, usize>>();
+        for grade in &self.grades {
+            r.entry(grade.grade).and_modify(|x| *x += 1);
+        }
+        r
+    }
+}
+
+//--------------------------------------------------------------------------------------------------
+
+struct Stat {
+    description: String,
+    value: String,
+}
+
+impl Stat {
+    fn new(description: &str, value: &str) -> Stat {
+        Stat {
+            description: description.to_string(),
+            value: value.to_string(),
+        }
+    }
+
+    fn markdown(&self) -> Vec<String> {
+        vec![self.description.clone(), self.value.clone()]
+    }
+}
+
+struct Stats {
+    stats: Vec<Stat>,
+}
+
+impl Stats {
+    fn new() -> Stats {
+        Stats { stats: vec![] }
+    }
+
+    fn push(&mut self, stat: Stat) {
+        self.stats.push(stat);
+    }
+
+    fn markdown(&self) -> String {
+        let mut rows = vec!["Description,Value"
+            .split(',')
+            .map(|x| x.to_string())
+            .collect::<Vec<_>>()];
+        rows.append(&mut self.stats.iter().map(|x| x.markdown()).collect::<Vec<_>>());
+        md_table(&rows, &[])
+    }
+}
+
+//--------------------------------------------------------------------------------------------------
+
+fn measure(rows: &[Vec<String>]) -> Vec<usize> {
+    let mut r = rows[0].iter().map(|_| 0).collect::<Vec<_>>();
+    for row in rows.iter() {
+        for (i, cell) in row.iter().enumerate() {
+            r[i] = r[i].max(cell.chars().collect::<Vec<_>>().len());
+        }
+    }
+    r
+}
+
+fn md_table(rows: &[Vec<String>], right: &[usize]) -> String {
+    let widths = measure(rows);
+    let mut r = vec![];
+    for (i, row) in rows.iter().enumerate() {
+        let mut s = vec![];
+        for (j, cell) in row.iter().enumerate() {
+            let sep = if j == 0 { "" } else { " | " };
+            s.push(if right.contains(&j) {
+                format!("{sep}{cell:>0$}", widths[j])
+            } else {
+                format!("{sep}{cell:<0$}", widths[j])
+            });
+        }
+        r.push(format!("| {} |\n", s.join("")));
+        if i == 0 {
+            let mut s = vec![];
+            for (j, _cell) in row.iter().enumerate() {
+                s.push(format!(
+                    "{}{}{}",
+                    if j == 0 { "" } else { "|-" },
+                    "-".repeat(widths[j]),
+                    if right.contains(&j) { ':' } else { '-' }
+                ));
+            }
+            r.push(format!("|-{}|\n", s.join("")));
+        }
+    }
+    r.join("")
 }