+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
 * 0.1.0 (2023-12-06): Initial release
-* 0.1.1 (2023-12-06): Save the quiz grading report to a file
-* 0.1.2 (2023-12-06): Fix typo
-* 0.1.3 (2023-12-07): Clean up readme and code
+    * 0.1.1 (2023-12-06): Save the quiz grading report to a file
+    * 0.1.2 (2023-12-06): Fix typo
+    * 0.1.3 (2023-12-07): Clean up readme and code
+* 0.2.0 (2023-12-07): Streamline design: `-g` is now `-q`, improved tables;
+  debug: bat pager, don't create quiz directories; fix typo